requirements.txt
bcrypt==4.2.1
blinker==1.8.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
Flask==3.0.3
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-DebugToolbar==0.16.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.0.3
WTForms==3.2.1
